+++ package.json
@@ -6,7 +6,7 @@
   "type": "module",
   "scripts": {
     "watch": "node watch_build.js",
-    "build": "esbuild src/js/scripts.js --minify --outfile=dist/js/scripts.min.js && esbuild src/js/cookies.js --minify --outfile=dist/js/cookies.min.js && esbuild src/css/style.css --minify --outfile=dist/css/style.min.css && esbuild src/css/home.css --minify --outfile=dist/css/home.min.css && esbuild src/js/highlight.js --minify --outfile=dist/js/highlight.min.js && esbuild src/js/print-holdings-checker.js --minify --outfile=dist/js/print-holdings-checker.min.js && esbuild src/css/print-holdings-checker.css --minify --outfile=dist/css/print-holdings-checker.min.css"
+    "build": "esbuild src/js/scripts.js --minify --outfile=dist/js/scripts.min.js && esbuild src/js/cookies.js --minify --outfile=dist/js/cookies.min.js && esbuild src/css/style.css --minify --outfile=dist/css/style.min.css && esbuild src/css/home.css --minify --outfile=dist/css/home.min.css && esbuild src/js/highlight.js --minify --outfile=dist/js/highlight.min.js && esbuild src/js/print-holdings-checker/index.js --bundle --minify --outfile=dist/js/print-holdings-checker.min.js && esbuild src/css/print-holdings-checker.css --minify --outfile=dist/css/print-holdings-checker.min.css"
   },
   "repository": {
     "type": "git",
